[protobuf] Update to protobuf 3.11.4
