+++ recipes/protobuf/all/test_package/addressbook.pb.cc.3.9.1
