+++ recipes/protobuf/all/test_package/addressbook.proto
