+++ recipes/protobuf/config.yml
@@ -1,4 +1,6 @@
 versions:
   3.9.1:
     folder: 3.9.x
+  3.11.4:
+    folder: 3.11.x
